+++ sigima/tests/image/imageobj_unit_test.py
@@ -104,7 +104,7 @@ def test_hdf5_image_io() -> None:
                 execenv.print(f"  Skipping {fname} (not implemented)")
                 continue
             # Save to HDF5
-            filename = osp.join(tmpdir, f"test_{osp.basename(fname)}.h5")
+            filename = osp.join(tmpdir, f"test_{osp.basename(fname)}.h5ima")
             sigima.io.write_image(filename, orig_image)
             execenv.print(f"  Saved {filename}")
             # Read back